Portail_Torus
Nice animation du portails tore maintenant
Changement d'objet aussi hehe
J'espere un jour trouver une fonction plus rapide et elegante que rand()
% UN_CHIFFRE

+++ Sources/DLL/Arbre/Noeuds/NoeudGenerateurBille.h
@@ -45,7 +45,7 @@ private:
 	bool inverserDirection_ { false };
 	float compteurAnimation_ { 0 };
 	float compteurAnimationBouger_{ 0 };
-	float distanceBouger_ { 0 };
+	double distanceBouger_ { 0 };
 	// Vu que le temps n'est pas constant, on dois remettre la table a sa position par defaut
 	glm::dvec3 positionPreDeplacement_{ 0 };	
 };
